Only send INFO logs and higher to the log channel
This aligns with what it printed to the console, even though much more is logged via `DEBUG` and `TRACE`. This also
avoids the hangs we were experiencing, since the volume of messages was so high that the create message requests were
being throttled and backed up. It took about 10-15 for these to clear, even after running the server for only a few
seconds.

This is accomplished by a change from a custom filter to a customer appender. The latter is the correct way to handle
log messages and send them to a target, while filters are for deciding which message to keep and which to drop.