Simply ConfigManager to only handle reading and writing config files
`LinkedProfile`s are now managed by a `LinkedProfileRepository`. While they are still stored under the `discordant/`
config directory, that can be considered and implementation detail and thus the `ConfigManager` does not care. In the
future, `ConfigManager` can manage a dedicated `discordant/config/` directory as we expand the number of configuration
files. The `LinkedProfile` `record` has to be converted back to a `class` as Gson was unable to set the fields, even
though it can set the `final` fields of a `class`... confusing.

`ConfigManager` is now simpler in that the number of public methods has been reduced. `craftPaths()` and
`genDiscordLinkSettings()` have been merged into a single method `ensureConfigStructure()`, and no support for linked
profiles remains. Instead of operating on a mixture of `Path`, `String` and `new File(...)`, we now always represent
paths as `Path`s and convert those to `File`s by `Path::toFile()`.