Revamp the message system
We now have a lot more consistency in message definitions. We support three types of messages:
- Discord messages are configured as keys in the "discord.messages" object, and support a title, color, thumbnail and
  description (all optional).
- Discord topics are configured as keys in the "discord.topics" object, and support a description.
- Minecraft messages are configured as keys in the "minecraft.messages" object, and support a text.

All three message types support scoped parameter substitution in all of their fields, with all fields of a given message
sharing the same scope.

After substitution:
- Discord topics are used to plainly set the chat channel topic.
- Discord messages are converted into embeds to add to the chat channel.
- Minecraft messages are converted into components to be sent to the clients.

Most of the visible changes are to Discord messages. As mentioned above, every discord message is now an embed. The
configured defaults are set to give - what I think is - a pleasant experience. For the most part, discord message will
be default have a title and a description. Player join messages with have the player head as a thumbnail. Player chat
messages with have the player as author and no title or thumbnail. Only player chat messages will have an author set -
the rest of the messages rely on the title to indicate their source.

Since we only use webhooks for discord messages (in addition to plain text for the log channel), we no longer need
webhook support. So that has been removed along with the Unirest dependency. `config.json` no longer has any webhook
fields in it.

Some other configuration defaults have been changed to improve the onboarding process. Mainly, we now assume that
account linking is desirable and so that is the default (it can still be disabled). We also make the assumption that
when account linking is enabled, we should always use it for rendering messages. As such, `"useDiscordData"` and
`"preloadDiscordMembers"` have been removed with the new behavious being as though they were both set to `true`.

We have kept it so that the same message representation (`SemanticMessage`) can be rendered as a Minecraft message or as
a Discord message while maintaining the semantics of the message - e.g., in Minecraft we can color, hover, click,
etc. But the representation of `SemanticMessage` has been changed so that it is essentially a list of substrings, with
each substring having an optional tag that describes its purpose and provides some metadata. Plaintext messages can now
be created by ignoring the tags and appending the substrings without needing to be updated whenever a new tag type is
added.

Player head URLs have not been integrated into the message system yet (i.e., there is no formal scope for them). But a
new parameter has been added (`time`) which will be replaced with the current unix time. This can be useful for dodging
Discord's excessive caching length as it doesn't respect the cache control of the original.