Shutdown JDA a little later in the server lifecycle
We used to shut down JDA when `SERVER_STOPPING` was fired. However, there are still logs being written that a server
admin may want to be sent to Discord. So instead we wait for `SERVER_STOPPED`, at which point there should be no (or
almost no) log messages still being written.

The chat channel topic and shutdown message are still sent on `SERVER_STOPPING`.