Bend JDA shutdown to our will
Instead of `.queue()`ing requests, we now `.submit()` them. Not only does `submit()` return a future (that we could
handle in the future), but it gives us an option via parameter to avoid the requeueing behaviour when a rate limit is
achieved. For now, we always pass `true` to indicate that requests should be requeued, but in the future we may change
this by dynamically disabling the requeue, or by being selective about which messages can be requeued.

We now have a two-stage shutdown process for JDA. If JDA shutdown can be actioned quickly enough (< 3 seconds), then the
server shuts down peacefully without further ado. But if JDA shutdown takes longer, we cancel all the requests and then
ensure JDA is shutdown.