Move DiscordListener and related message parsers into core
In order to help out the move, we have defined a new `Profile` interface that abstracts over a `GameProfile`, providing
just a UUID and name without the extra functionality of a `Player`. This is also what is used in `PlayerScope` now as
that is all the information we need and all the information that the player-join-attempt event can provide.

The message parsers have had their names tweaked to read a little better, and operate by building a `SemanticMessage`
rather than a `Component`.