Move non-mixin mc classes out of their subpackages
Too many of these classes were lone classes inside their respective subpackages. That is just unnecessary, so we moved
them out. The exception is the `mixin` package, which holds the mixins. Not only are there multiple mixins, but Fabric
expects mixins to be in a separate package, so our hands are bound there.

We also removed `ServerLanguage` which was responsible for loading translations from mods. This just doesn't have
anything to do with interfacing with Discord, so another mod should provide this functionality.