Simplify scope definitions
The `Scope` interface no longer uses CRTP as that wasn't providing any value - it was a holdover from the first take on
scopes. Additionally, we now have a child interface `DerivedScope` that has default behaviour for merging a scope's
values with the values from its derived scope, meaning each scope can focus on defining what matters to it. the
`Scope.parameters()` method has been removed as we aren't doing pre-parsing anyways, and just-in-time validation can be
done with only the parameter map.

Scopes no longer rely on anything from `modinterfaces`. While convenient for the scope, and somewhat convenient for
instantiating the scope, it got in the way of creating scopes in the absence of references to the server, players and
other objects. Instead, each parameter value must be provided to the scope's constructor. The result is the scopes are
purely data-oriented and the mssaging system is more self-sufficient.

All scopes now consistently prefix their parameters. E.g., all `PlayerScope` parameter names start with `"player."`. The
only quirk is that `ProfileScope` and `PlayerScope` share a prefix, just because `{player.name}` and `{player.uuid}` is
more intuitive that `{profile.name}` and `{profile.uuid}` when other player-specific things would use the `"player"`
prefix.

Avatar and icon URLs now use real scopes, specifically the new `ProfileScope`. This way they can support more than the
three basic parameters of `{player.uuid}`, `{player.name}` and `{server.time}` and they can be documented in the same
way as other parameterized strings. In order to make this work, we added a new `ProfileScope` that only holds a player
name and UUID, but not any discord user or avatar/icon URLs (without this, we have infinite recursion trying to
instantiate the URL templates). `PlayerScope` now inherits from `ProfileScope` in order to make sure we don't have a
discrepency between player name and UUId parameters.