Simplify scope definitions
The `Scope` interface no longer uses CRTP as that wasn't providing any value - it was a holdover from the first take on
scopes. Nor do we have much in the way of helper functionality in `Scope` anymore, nor do we have subinterfaces to help
out. Instead, `Scope` is built entirely around one method (`addValuesTo()`) which can add parameters to an immutable map
builder. This has many benefits, including extreme simplicity and the fact that the builder checks that no duplicate
keys exist. A convenience method (`values()`) creates a builder to pass to `addValuesTo()` and then returns the map in
the end.

`Scope` no longer has a `parameters()` method in `Scope` because we aren't doing pre-parsing anyways. Just-in-time
validation could be done in the future based solely on the keys returned by `values()`.

Avatar and icon URLs now use real scopes, specifically the new `ProfileScope`. This way they can support more than the
three basic parameters of `{player.uuid}`, `{player.name}` and `{server.time}` and they can be documented in the same
way as other parameterized strings. In order to make this work, we added a new `ProfileScope` that only holds a player
name and UUID, but not any discord user or avatar/icon URLs (without this, we have infinite recursion trying to
instantiate the URL templates). `PlayerScope` now inherits from `ProfileScope` in order to make sure we don't have a
discrepency between player name and UUId parameters.

Scopes no longer rely on anything from `modinterfaces`. While convenient for the scope, and somewhat convenient for
instantiating the scope, it got in the way of creating scopes in the absence of references to the server, players and
other objects. Instead, each parameter value must be provided to the scope's constructor. The result is the scopes are
purely data-oriented and the mssaging system is more self-sufficient. Tedious cases (e.g., instantiated `ServerScope`
and `PlayerScope`) can be mitigated by helpers.

All scopes now consistently prefix their parameters. E.g., all `PlayerScope` parameter names start with `"player."`. The
only quirk is that `ProfileScope`, `PlayerScope` and `DiscordUserScope` share a prefix, partly by historical accident
and partly because `{player.name}` and `{player.uuid}` is more intuitive for an end user than `{profile.name}` and
`{profile.uuid}` when other player-specific things would use the `"player"` prefix.