Fix okio version
The okio version was too high (>3.0) when okhttp was expecting 2.8.0. This caused okio to not be included via JAR-in-JAR
and resulted errors on servers. The development environment was fine as the true dependency was locally available.

We also added an explicit dependency on the kotlin standard library, because it is required by okio and okhttp as well.