Replace integer authentication codes with structured tokens
The new tokens are composed of a player UUID and an authentication code. The code is in the same format as the previous
authentication codes, which is a six-digit randomly generated integer. This makes a token strongly associated with a
specific player and means we can more easily action a verification request as we always have player data available.

This cleans up `Discordant` as we only have to maintain a single hash table for pending verifications, as opposed to two
that needed to be kept in sync in the previous implementation.