Allow reloading configurations
This is a very involved change because it doubles as some serious cleanup and restructuring. The final result is that
the main `Discordant` class is built from relatively few components, most of which do not need to react to configuration
changes. A handful of these components do need to be updated when configurations are reloaded, but they have been
designed in a way that the reload can be handled internally to them. Importantly, none of these components need to be
recreated on reload.

The few components that need to be reloaded are:
- `DiscordApi`, so old connections can be closed and new connections created.
- `LinkedProfileManager`, so linking requirements and timeouts can be updated.
- `ScopeFactory`, so the server name and URL templates can be updated.
- `DiscordantMessageHandler`, so messages can be reloaded, and linking and mention configs can be updated.

The `LinkedProfileManager` is a new component that has been added to take full responsibility for managing linked
profiles. It complements `LinkedProfileRepository` - which handles storing linked profiles - by handling pending links,
verification of links, and provides lookups of player names by discord user ID.

In order to support the breakdown of components, we have split out the `EmbedFactory` and `ScopeFactory` for creating
embed messages and instantiating message templates, respectively. Similarly, a `TickedClock` now keeps track of time. No
longer do any components have to depend on `Discordant` itself.

The `/discord reload` command has been added, which attempts to load a new configuration. If successful, it will inject
that new configuration into the various components that need to be updated.