Improve Discord -> Minecraft messages and mentions
The use of `!+` to directly mention a Minecraft player has been removed. Instead only Discord users can be mentioned,
which is done via the usual Discord syntax. For any Discord user linked to a Minecraft player, the user mention will be
replaced with the player name, falling back to the Discord nickname.

In addition to the sender, each player/user mentioned in the message will be coloured according to the corresponding
Discord role. In addition, each mentioned can be hovered over in Minecraft to show the role name and user tag of the
Discord user. Clicking will now not only copy the user tag, but a leading `@` as well, so pasting into Discord will
automatically create a mention of the user.

A new utility method - `StringUtil.chunk()` - has been added in order to carve a string into multiple parts according to
a regular expression. It is similar to splitting a string, but will return both matched and unmatched parts so they can
both be processed. The new messaging functionality is built on top of this utility, and `ScopeUtil.instantiate()` has
also been migrated to it.