Handle role and channel mentions in Discord messages
Roles are now rendered similar to user name by being italicized and coloured according the role.  But there is no logic to
figure out which players to notify based on role.

Channel mentions are also handled and are rendered as simple underlined text. Clicking on them will open the jump URL in
a browser.